refactor(all): file paths changed in comply with change in the file names.
