+++ src/application/services/loan-service.ts
@@ -1,8 +1,8 @@
 import { Loan } from "../../domain/entities/loan.js";
-import { ILoanRepository } from "../../domain/repositories/i-loan.repository.js";
-import { IBookRepository } from "../../domain/repositories/i-book.repository.js";
-import { IUserRepository } from "../../domain/repositories/i-user.repository.js";
-import { BaseService } from "./base.service.js";
+import { ILoanRepository } from "../../domain/repositories/i-loan-repository.js";
+import { IBookRepository } from "../../domain/repositories/i-book-repository.js";
+import { IUserRepository } from "../../domain/repositories/i-user-repository.js";
+import { BaseService } from "./base-service.js";
 
 export type CreateLoanDto = Pick<Loan, 'userId' | 'bookId'>
 export type UpdateLoanDto = Partial<CreateLoanDto>;
