+++ src/domain/repositories/i-book-repository.ts
@@ -1,5 +1,5 @@
 import { Book } from '../entities/book.js'
-import { IGenericRepository } from './i-generic.repository.js';
+import { IGenericRepository } from './i-generic-repository.js';
 
 export interface IBookRepository  extends IGenericRepository<Book>{
